Change the mine generation algorithm
If on first pick, you tap over a mine, the new algorithm will move the
bomb to the upper left corner and recompute the numbers. This is similar
to what Windows does (can be checked by using the XYZZY trick).
Previously I generated the board on first tap, but this is not what
Windows does.